proxy-vcr/pyproject.toml
[build-system]
requires = ["uv_build>=0.8.10"]
build-backend = "uv_build"

[project]
name = "proxy-vcr"
version = "0.1.0"
description = "The Proxy VCR is a tool that records and replays HTTP requests."
readme = "README.md"
dependencies = [
    "google-auth>=2.40.3",
    "httptools>=0.6.4",
    "httpx>=0.28.1",
    "openai>=1.99.9",
    "pydantic-settings>=2.10.1",
    "rich>=14.1.0",
    "starlette>=0.47.2",
    "uvicorn>=0.35.0",
    "uvloop>=0.21.0",
    "vcrpy>=7.0.0",
    "watchfiles>=1.1.0",
]

[tool.uv.build-backend]
module-name = "proxy_vcr"
module-root = ""
